requirements.txt
psycopg2>=2.9.3
python-dotenv>=0.19.2
requests>=2.28.0
pyhive[hive]>=0.6.5
thrift>=0.16.0
sasl>=0.3.1
thrift-sasl>=0.4.3